pyproject.toml
[tool.poetry]
name = "cls-forwarding-service"
version = "0.1.0"
description = "Transfer files from on-prem fileshares to cloud"
authors = ["Laurent Lejeune <laurent.lejeune@bayer.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.3"
watchdog = "^2.2.1"
boto3 = "^1.26.73"
click = "^8.1.3"
SQLAlchemy = "^2.0.3"
hvac = "^1.0.2"
PyYAML = "^6.0"
gunicorn = "^20.1.0"
marshmallow = "^3.19.0"
flask-smorest = "^0.40.0"
flask-sqlalchemy = "^3.0.3"
prettytable = "^3.6.0"
flask-rq2 = "^18.3"
sqlalchemy-utils = "^0.40.0"
marshmallow-sqlalchemy = "^0.29.0"
flask-marshmallow = "^0.14.0"
rich = "^13.3.2"
python-dotenv = "^1.0.0"
python-decouple = "^3.8"
flask-migrate = "^4.0.4"

[tool.poetry.group.dev.dependencies]
yapf = "*"
isort = "^5.2.2"
mypy = "^0.782"
pytest-flask = "^1.2.0"
pyright = "*"
black = "^22.3.0"
python-language-server = {extras = ["all"], version = "^0.36.2"}
moto = {extras = ["all"], version = "^4.1.3"}
mock-file-tree = "^1.1.3"
responses = "^0.22.0"
flask-cors = "^3.0.10"
sure = "^2.0.1"
pytest-mock = "^3.10.0"
aws-error-utils = "^2.7.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
